changes from review, codeql
