+++ docs/api/utils.rst
@@ -1,0 +1,15 @@
+.. _utils:
+
+Utility functions
+=================
+
+RosettaSciIO provides certain utility functions that are applicable for multiple
+formats, e.g. for the HDF5-format on which a number of plugins are based.
+
+.. _hdf5-utils:
+
+HDF5 utility functions
+^^^^^^^^^^^^^^^^^^^^^^
+
+.. automodule:: rsciio.utils.hdf5
+   :members:
