+++ docs/file_specification/index.rst
@@ -1,8 +1,13 @@
 File Specifications
 ===================
 
-This section contains reference documentation for file specification available
+This section contains reference documentation for file specifications available
 publicly.
+
+.. toctree::
+   :hidden:
+   
+   ripple-specs
 
 .. _edax-file_specification:
 
@@ -15,3 +20,10 @@ available publicly available from EDAX:
 - :download:`spc <edax/SPECTRUM-V70.pdf>`
 - :download:`spd <edax/SpcMap-spd.file.format.pdf>`
 - :download:`ipr <edax/ImageIPR.pdf>`
+
+
+Ripple (Lispix)
+---------------
+
+:ref:`Format description <ripple-file_specification>` for the :ref:`ripple
+<ripple-format>` file format.
