+++ docs/supported_formats/netcdf.rst
@@ -9,8 +9,9 @@ but we do not support writing to this format.
 
 Note that only NetCDF files written by EELSLab are supported.
 
-To use this format a python netcdf interface must be installed manually because
-it is not installed by default when using the automatic installers.
+To use this format a python netcdf interface, such as `netcdf4
+<http://unidata.github.io/netcdf4-python/>`_ must be installed manually because
+it is not installed by default.
 
 
 API functions
