+++ rsciio/nexus/__init__.py
@@ -1,16 +1,12 @@
 from ._api import (
     file_reader,
     file_writer,
-    list_datasets_in_file,
-    read_metadata_from_file,
 )
 
 
 __all__ = [
     "file_reader",
     "file_writer",
-    "list_datasets_in_file",
-    "read_metadata_from_file",
 ]
